requirements.txt
openai
python-dotenv
google-generativeai
Flask